# admin/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabio POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outlets/{outletId}/orders/{orderId}/kot:
    post:
      operationId: generateKot
      tags: [kot]
      summary: Generate KOT
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KOT generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KOT"
  /outlets/{outletId}/kots:
    get:
      operationId: getKots
      tags: [kot]
      summary: List KOTs
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: KOTs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KOT"
  /outlets/{outletId}/kots/{kotId}:
    put:
      operationId: updateKot
      tags: [kot]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: kotId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKotRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KOT"

  # INVENTORY
  /outlets/{outletId}/inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /outlets/{outletId}/inventory/{itemId}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /outlets/{outletId}/inventory/{itemId}/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      parameters:
        - name: outletId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryRequest"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # REPORTS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        variantId:
          type: integer
          nullable: true
        variantName:
          type: string
          nullable: true
        addonIds:
          type: array
          items:
            type: integer
        addonNames:
          type: array
          items:
            type: string
        addonTotal:
          type: number
        totalPrice:
          type: number
        notes:
          type: string
          nullable: true
        kotStatus:
          type: string
          enum: [pending, preparing, ready, served]
      required: [id, menuItemId, menuItemName, quantity, unitPrice, addonIds, addonNames, addonTotal, totalPrice, kotStatus]
...
    KOT:
      type: object
      properties:
        id:
          type: integer
        outletId:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: string
        kotNumber:
          type: string
        tableId:
          type: integer
          nullable: true
        tableName:
          type: string
          nullable: true
        orderType:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          type: string
          enum: [pending, preparing, ready, served]
        station:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, outletId, orderId, orderNumber, kotNumber, orderType, items, status, station, createdAt, updatedAt]
    UpdateKotRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, preparing, ready, served]
      required: [status]
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        outletId:
          type: integer
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        maxStock:
          type: number
        costPerUnit:
          type: number
        isLowStock:
          type: boolean
        lastUpdated:
          type: string
      required: [id, outletId, name, category, unit, currentStock, minStock, maxStock, costPerUnit, isLowStock, lastUpdated]
    CreateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        maxStock:
          type: number
        costPerUnit:
          type: number
      required: [name, category, unit, currentStock, minStock, maxStock, costPerUnit]
    AdjustInventoryRequest:
      type: object
      properties:
        adjustment:
          type: number
        reason:
          type: string
      required: [adjustment, reason]
...